Создал отдельный класс сингтон для контекста, подправил отступы, сдаю работу
